+++ example/CountlyRNExample/DeviceID.tsx
@@ -6,11 +6,11 @@ import CountlyButton from "./CountlyButton";
 import { lightOrange } from "./Constants";
 
 const temporaryDeviceIdMode = () => {
-    Countly.changeDeviceId(Countly.TemporaryDeviceIDString, true);
+    Countly.deviceId.changeDeviceId(Countly.TemporaryDeviceIDString, true);
 };
 
 const changeDeviceId = () => {
-    Countly.changeDeviceId("02d56d66-6a39-482d-aff0-d14e4d5e5fda", true);
+    Countly.deviceId.changeDeviceId("02d56d66-6a39-482d-aff0-d14e4d5e5fda", true);
 };
 
 function DeviceIDScreen({ navigation }) {